Linux crash handling

+++ terrarium-core/include/terrarium/defines.hpp
@@ -41,19 +41,19 @@ namespace terra::inline primitives {
 }
 
 namespace terra::inline literals::inline memory_literals {
-    [[nodiscard]] consteval auto operator""_b(const usize size) noexcept -> usize {
+    [[nodiscard]] consteval auto operator""_b(const unsigned long long size) noexcept -> usize {
         return size;
     }
 
-    [[nodiscard]] consteval auto operator""_kb(const usize size) noexcept -> usize {
+    [[nodiscard]] consteval auto operator""_kb(const unsigned long long size) noexcept -> usize {
         return size * 1024U;
     }
 
-    [[nodiscard]] consteval auto operator""_mb(const usize size) noexcept -> usize {
+    [[nodiscard]] consteval auto operator""_mb(const unsigned long long size) noexcept -> usize {
         return size * 1024U * 1024U;
     }
 
-    [[nodiscard]] consteval auto operator""_gb(const usize size) noexcept -> usize {
+    [[nodiscard]] consteval auto operator""_gb(const unsigned long long size) noexcept -> usize {
         return size * 1024U * 1024U * 1024U;
     }
 }
